test: the utils tests follow their modules
test-utils.js splits into test-hex.js, test-binary.js, test-debounce.js and
test-keymap.js; the zip and gzip tests join the crc and zip writer tests in
test-archive.js, one file per module. The README points at the key tables'
new homes.
